pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "slateup-funding"
version = "1.0.0"
description = "SlateUp Funding API with BoldSign eSignature integration"
requires-python = ">=3.10"
dependencies = [
    "boldsign>=3.0.0",
    "email-validator>=2.0.0",
    "fastapi>=0.109.0",
    "requests>=2.28.0",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.27.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.23.0", "httpx>=0.27.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "backend"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]
